requirements.txt
absl-py==2.1.0
accelerate==1.5.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albumentations==1.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
blis==1.2.0
Brotli==1.1.0
captum==0.7.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.3.1
ctranslate2==4.5.0
cycler==0.12.1
cymem==2.0.11
datasets==3.4.1
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastprogress==1.0.3
fasttext-wheel==0.9.2
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.0
Flask-Compress==1.17
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gensim==4.3.3
gradio==5.33.2
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.4.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
OpenNMT-py==3.5.1
optree==0.14.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.1.1
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pyarrow==19.0.1
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyonmttok==1.37.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qudida==0.0.4
RapidFuzz==3.12.2
rdkit==2024.9.6
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.11.13
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selfies==2.2.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
SmilesPE==0.0.3
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.13
timm==0.4.12
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.0.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.3
waitress==3.0.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zstandard==0.23.0
